+++ modelserver/types/locator.py
@@ -8,14 +8,14 @@ class HFLocator(BaseModel):
     Locator type for validating the different set of possible revisions.
     """
 
-    type: Literal["hf"] = "hf"
+    type: Literal["locatorv1/hf"] = "locatorv1/hf"
     repo: str
     file: str
     revision: str | None = None
 
 
 class DiskLocator(BaseModel):
-    type: Literal["disk"] = "disk"
+    type: Literal["locatorv1/disk"] = "locatorv1/disk"
     path: str
